Updated Pages
Updated sign up page, login and payment page to reference SSN for the user login, sign up and payments.

Updated payment page, originally it was sending a default transaction ID of 5. The SQL table was built to auto increment the transaction ids by default. Not necessary to include a query for the transaction.

Added a home logo to a new 'images' folder. However, added the image to main adds it to all the remaining pages. - Still troubleshooting.